Major update separating code from config and layout.
Config is no longer part of the launcher itself, neither is the layout.
Two configs and layouts are shipped with this version; a default and an MPV example.
Commands are now non-blocking, so you can now continue to control the launcher while MPV is launched in the background for instance.
Padding is calculated a bit better now too, allowing for odd numbers.
Theoretically buttons can have their vertical padding specified too, though it's not yet part of the Config.